+++ lib/src/ff_auto_import.dart
@@ -1,0 +1,9 @@
+/// Annotation of auto import
+class FFAutoImport {
+  const FFAutoImport([this.suffix]);
+
+  /// The suffix of this import
+  /// for example, maybe you need to hide,show
+  /// hide xyz,show abc
+  final String? suffix;
+}